# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Shakai Public API
  version: 1.0.0
  description: 外部クライアント向けのバージョン付きREST API
servers:
  - url: https://{host}/api/v1
    variables:
      host:
        default: example.com
security:
  - bearerAuth: []
paths:
  /me:
    get:
      summary: 現在のユーザーを取得
      responses:
        '200':
          description: User
        '401':
          $ref: '#/components/responses/Unauthorized'
  /posts:
    get:
      summary: 投稿一覧を取得
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        '200': { description: Feed items }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      summary: 投稿を作成
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostInput'
      responses:
        '201': { description: Created }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /me/profile:
    get:
      summary: 自分のプロフィールを取得
      responses:
        '200': { description: Profile }
        '401': { $ref: '#/components/responses/Unauthorized' }
    patch:
      summary: 自分のプロフィールを更新
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProfileInput' }
      responses:
        '200': { description: Updated profile }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /profiles/{handle}:
    get:
      summary: プロフィールを取得
      parameters:
        - in: path
          name: handle
          required: true
          schema: { type: string }
      responses:
        '200': { description: Profile }
        '404': { description: Profile not found }
  /profiles/{handle}/follow:
    post:
      summary: プロフィールをフォロー
      parameters:
        - in: path
          name: handle
          required: true
          schema: { type: string }
      responses:
        '200': { description: Following state }
        '401': { $ref: '#/components/responses/Unauthorized' }
    delete:
      summary: プロフィールのフォローを解除
      parameters:
        - in: path
          name: handle
          required: true
          schema: { type: string }
      responses:
        '200': { description: Following state }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /profiles/suggestions:
    get:
      summary: おすすめプロフィールを取得
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 20, default: 3 }
      responses:
        '200': { description: Suggested profiles }
        '401': { $ref: '#/components/responses/Unauthorized' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Better-Auth-session-token
  schemas:
    ProfileInput:
      type: object
      properties:
        bio: { type: string, maxLength: 400 }
        currentThought: { type: string, maxLength: 400 }
        interests:
          type: array
          items: { type: string }
    PostInput:
      type: object
      required: [genre, division, form, title, body]
      properties:
        genre: { type: string }
        division: { type: string }
        form: { type: string }
        title: { type: string }
        body: { type: string }
    Error:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationError:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }